Removed dependency on any one libnotify client
The problem was when using python2 or python3,
there were a lot of problems when the said modules
for notification weren't supported by either of the
python version.

This commit fixes this problem, by importing the
appropriate module based on the python version
used.

This changes a way of running the script. You should
now run this script as:

python scorer.py