Restructured the code, removed some hardcoding
This is just a temperory fix. I had to remove extra
features like getting the latest overs and other
information because many times the json was
complaining for KeyError. So I need we need to understand
the structure of json in great detail.

So now the structure is defined like this
scorer:
	__init__.py
	__main__.py
	app.py
	fetch_scores.py
	notification.py

You need to run the script as `python app.py`

Most of the functionality is implemented in the fetch_scores.py but I
am not using those functions for now.

I will work more on restructing the code in the next few commits,
so I need to make use of __init__.py and __main__.py

Only then this branch will be merged with master.