Migrated the project to Python version 2.
As the project was increasing, there were lots of problems to maintain
the code with respect to both python 2 as well as python 3. There were
bugs being introduced in the project everytime a new code base was
introduced in the project. And some of the developers were still using
python 2.

Much of this decision was also influened with pynotify. Since the
package is widely available for python 2, and not for python 3. The
other packages that are there for python 3 are not consistent across
platforms with their function names, cause even more problems for making
the project support as mnay linux based OS as possible.

Also there were a lot of problems packaging the project according to the
standards of distutils/setuptools.

Henceforth all the developement will be done with python 2 only.