Used a single instance of logger globally
I have used some techniques in the python logging cookbook[1]
to make changes to the logging mechanism so far.

I have also used some guidelines mentioned in the official logging
HOWTO[2] on when to use logging.info and when to use logging.debug and
made changes to the code accordingly.

[1]https://docs.python.org/2/howto/logging-cookbook.html
[2]https://docs.python.org/2/howto/logging.html